make the space in the beginning of phrase belong to the phrase

+++ src/phrase.cpp
@@ -8,7 +8,7 @@ Phrase::Phrase(uint64_t id, const std::string &phrase, const std::string &transl
 
 }
 
-size_t Phrase::size() const
+int64_t Phrase::size() const
 {
     return phrase.size();
 }
